Refactor UserDataStore to UserLocalDataStore
