--- data/src/main/java/in/koreatech/koin/data/di/datastore/DataStoreModule.kt
@@ -1,31 +1,0 @@
-package `in`.koreatech.koin.data.di.datastore
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStore
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import `in`.koreatech.koin.data.source.local.UserLocalDataSource
-import `in`.koreatech.koin.data.source.datastore.UserDataStore
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DataStoreModule {
-
-    private val Context.userDataStore: DataStore<Preferences> by preferencesDataStore(
-        name = UserLocalDataSource.PREF_NAME
-    )
-
-    @Provides
-    @Singleton
-    fun provideUserDataStore(
-        @ApplicationContext context: Context
-    ): UserDataStore {
-        return UserDataStore(context.userDataStore)
-    }
-}
